backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
jinja2==3.1.2
chromadb==0.5.0
sentence-transformers==2.7.0
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
httpx==0.25.2
numpy<2.0
PyPDF2==3.0.1
python-docx==1.1.0
python-magic==0.4.27